# shared/contracts/openapi/pokefolio-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PokeFolio API",
    "version": "1.0.0-alpha.1",
    "description": "Versioned contract for authenticated PokeFolio clients. Private ownership is always derived from the authenticated principal."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/collection": {
      "get": {
        "operationId": "listCollection",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Page of holdings owned by the authenticated account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "operationId": "createHolding",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHoldingCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Holding created for the authenticated account",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionHolding"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/collection/{holdingId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/HoldingId"
        }
      ],
      "get": {
        "operationId": "getHolding",
        "responses": {
          "200": {
            "description": "Owned holding",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionHolding"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "patch": {
        "operationId": "updateHolding",
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHoldingCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated holding",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionHolding"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "412": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "delete": {
        "operationId": "deleteHolding",
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "204": {
            "description": "Holding tombstoned"
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "412": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/collection/{holdingId}/quantity-delta": {
      "parameters": [
        {
          "$ref": "#/components/parameters/HoldingId"
        }
      ],
      "post": {
        "operationId": "applyQuantityDelta",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityDeltaCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Delta committed atomically and idempotently",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionHolding"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "HoldingId": {
        "name": "holdingId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": ["string", "null"],
          "maxLength": 512
        }
      },
      "PageSize": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 100
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Stable mutation UUID. For quantity-delta commands it must equal the body operationId.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "One strong holding ETag returned by the API, for example \"v3\". Wildcards, weak validators and lists are rejected.",
        "schema": {
          "type": "string",
          "pattern": "^\\\"v[1-9][0-9]*\\\"$",
          "maxLength": 22
        }
      }
    },
    "headers": {
      "ETag": {
        "description": "Opaque optimistic-concurrency version",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "Machine-readable error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    },
    "schemas": {
...
      "CreateHoldingCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "cardId", "language", "variant", "condition", "quantity"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cardId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "language": {
            "$ref": "#/components/schemas/CardLanguage"
          },
          "variant": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "condition": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000
          },
          "notes": {
            "type": ["string", "null"],
            "maxLength": 10000
          }
        }
      },
      "UpdateHoldingCommand": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "language": {
            "$ref": "#/components/schemas/CardLanguage"
          },
          "variant": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "condition": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "notes": {
            "type": ["string", "null"],
            "maxLength": 10000
          }
        }
      },
      "QuantityDeltaCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": ["operationId", "delta"],
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "delta": {
            "type": "integer",
            "minimum": -10000,
            "maximum": 10000,
            "description": "Non-zero signed delta applied atomically; zero is rejected by application validation."
          }
        }
      },
      "CollectionHolding": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "cardId", "language", "variant", "condition", "quantity", "version", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cardId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "language": {
            "$ref": "#/components/schemas/CardLanguage"
          },
          "variant": {
            "type": "string"
          },
          "condition": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0
          },
          "notes": {
            "type": ["string", "null"]
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CollectionPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "nextCursor"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionHolding"
            }
          },
          "nextCursor": {
            "type": ["string", "null"]
          }
        }
      },
...
      "CardLanguage": {
        "type": "string",
        "enum": ["de", "en", "ja", "fr", "it", "es", "ko", "zh-Hans", "zh-Hant"]
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": true,
        "required": ["type", "title", "status", "code", "correlationId"],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": ["string", "null"]
          },
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          }
        }
      }
    }
  }
}